Add files via upload
This project is developed as a part of Lets Upgrade Assignments. This app takes input from users and encrypts it, and saves the key. This key can be used to decrypt the ciphered text.
We have used the cryptography package of python and Fernet module. We create a file using the Genratekey function of fernet and then save its key, further used for encryption and decryption. Then we use genContent function to take the message that the user wants to decrypt. Now we can use the encrypt function for encryption and decrypt function for decryption of the file.